backend/requirements.txt
fastapi
uvicorn[standard]

python-dotenv

aiohttp
python-magic